Improves the numerical precision of the amplification factor (amplifiedChance).
This pull request improves the numerical precision of the amplification factor (amplifiedChance) used in passive skill calculations (such as dodge, momentum, transcendence).

+++ src/creatures/players/player.hpp
@@ -209,7 +209,7 @@ public:
 	void dismount();
 
 	uint16_t getDodgeChance() const;
-	uint16_t getAmplifiedChance() const;
+	double getAmplifiedChance() const;
 
 	uint8_t isRandomMounted() const;
 	void setRandomMount(uint8_t isMountRandomized);
